[holdings] Extract Account::State to its own file
