+++ lib/fortuneteller.rb
@@ -19,6 +19,7 @@ require 'fortuneteller/utils/social_security'
   require "fortuneteller/#{object_type}/generator"
   require "fortuneteller/#{object_type}/plan"
   require "fortuneteller/#{object_type}/transform" if File.exist?("lib/fortuneteller/#{object_type}/transform.rb")
+  require "fortuneteller/#{object_type}/state" if File.exist?("lib/fortuneteller/#{object_type}/state.rb")
 end
 require "fortuneteller/base/plan_reader"
